gobject-introspection.bbclass: Disable writing to $HOME/.cache/g-ir-scanner
Investigation of an end user going over their quota in the user's home
directory showed that the gobject-instrospection is writing files into
the ~/.cache/g-ir-scanner when building -native packages.  The test
package in this case was gdk-pixbuf-native.

Setting the GI_SCANNER_DISABLE_CACHE in the bbclass should be all that
is needed.  The qemu wrapper was already setting this properly, but we
also need it for the -native builds.

meta/recipes-gnome/gobject-introspection/gobject-introspection_1.58.3.bb:131:
export GI_SCANNER_DISABLE_CACHE=1

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>